Functional Kicking
Able to create from scratch: a field of view that lets the player know what target is available to kick. If they are within range, it will highlight the target and if the player hits the attack button, the target will have an implied force at the direction the player is facing the target.